Implement a `meteor shell` command.
Summary:
Invoking `meteor shell` or `meteor run --shell` now starts an interactive
REPL for evaluating server-side code instead of running the server.

The shell technically runs in the server process, so input and output have
to be piped from the parent process (i.e. the process that runs the
`meteor` command). This separation accounts for most of the complexity of
this diff, but the end result is pretty seamless: tab completion, history,
ctlr-C and ctrl-D, etc. all work as expected.

I have a list of remaining TODOs in the help.txt file.

Task: https://app.asana.com/0/15750483766338/16576093991355

Test Plan:
Run `meteor shell`, evaluate some commands, and verify that the commands
seem to be running the context of the server process. Also check tab
completion and history (TODO: carry history over between sessions), and
ensure that the server can be run simultaneously with the shell in a
separate terminal (TODO).

Reviewers: dgreenspan, avital, slava, emily, nim

Differential Revision: https://phabricator.meteor.com/D837